bugs at adding/removing criterium

+++ src/main/java/AHPSolver/XMLCreatorLogic.java
@@ -86,9 +86,9 @@ public class XMLCreatorLogic {
 		Element eCriterium = doc.createElement("criteria");
 		eCriterium.setAttribute("name", c.getName());
 		
-		for (Entry<String, Integer> entry : c.getValues().entrySet()) {
+		for (Entry<String, Double> entry : c.getValues().entrySet()) {
 			String key = entry.getKey();
-			Integer value = entry.getValue();
+			Double value = entry.getValue();
 			eCriterium.setAttribute(key, String.valueOf(value));
 		}
 		return eCriterium;
